improve double tab to zoom code, fix potential crash problem
