+++ samples/zoomable/src/main/java/com/facebook/samples/zoomable/ZoomableDraweeView.java
@@ -18,7 +18,6 @@ import android.content.Context;
 import android.content.res.Resources;
 import android.graphics.Canvas;
 import android.graphics.Matrix;
-import android.graphics.PointF;
 import android.graphics.RectF;
 import android.graphics.drawable.Animatable;
 import android.util.AttributeSet;
@@ -116,33 +115,6 @@ public class ZoomableDraweeView extends DraweeView<GenericDraweeHierarchy> {
     mZoomableController = createZoomableController();
     mZoomableController.setListener(mZoomableListener);
     mTapGestureDetector = new GestureDetector(getContext(), mTapListenerWrapper);
-    mTapGestureDetector.setOnDoubleTapListener(new GestureDetector.OnDoubleTapListener() {
-      @Override
-      public boolean onSingleTapConfirmed(MotionEvent e) {
-        return false;
-      }
-
-      @Override
-      public boolean onDoubleTap(MotionEvent e) {
-        float maxScaleFactor = ((DefaultZoomableController) mZoomableController).getMaxScaleFactor();
-        RectF bounds = ((DefaultZoomableController) mZoomableController).getImageBounds();
-        float x = e.getX();
-        float y = e.getY();
-        PointF imagePoint = new PointF((x - bounds.left) / bounds.width(), (y - bounds.top) / bounds.height());
-        float scaleFactor = mZoomableController.getScaleFactor() < maxScaleFactor ? maxScaleFactor : ((DefaultZoomableController) mZoomableController).getMinScaleFactor();
-        if (mZoomableController instanceof AbstractAnimatedZoomableController) {
-          ((AbstractAnimatedZoomableController) mZoomableController).zoomToPoint(scaleFactor, imagePoint, new PointF(x, y), DefaultZoomableController.LIMIT_ALL, 200, null);
-        } else {
-          ((DefaultZoomableController) mZoomableController).zoomToPoint(scaleFactor, imagePoint, new PointF(x, y));
-        }
-        return true;
-      }
-
-      @Override
-      public boolean onDoubleTapEvent(MotionEvent e) {
-        return false;
-      }
-    });
   }
 
   /**
@@ -200,6 +172,14 @@ public class ZoomableDraweeView extends DraweeView<GenericDraweeHierarchy> {
    */
   public void setTapListener(GestureDetector.SimpleOnGestureListener tapListener) {
     mTapListenerWrapper.setListener(tapListener);
+  }
+
+  /**
+   * Sets double tap listener for zoom or turn back, just support DefaultZoomableController
+   * and it's descendant currently.
+   */
+  public void setZoomDoubleTabListener(ZoomDoubleTabListener zoomDoubleTabListener) {
+    mTapGestureDetector.setOnDoubleTapListener(zoomDoubleTabListener);
   }
 
   /**
